requirements-langchain.txt
langchain>=0.0.267
langchain-community>=0.0.1
langchain-core>=0.0.1
openai>=0.27.8
google-search-results>=2.4.2  # For SerpAPI
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
